requirements.txt
# Movie Ticket Booking System - Backend Requirements
# Django Backend with PostgreSQL

Django>=5.0,<7.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
